Reverse back.

+++ database/migrations/2019_12_21_124911_article_category.php
@@ -15,9 +15,9 @@ class ArticleCategory extends Migration
     {
         Schema::create('article_category', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('article_id')->unsigned();
+            $table->integer('article_id')->unsigned();
             $table->foreign('article_id')->references('id')->on('articles')->onDelete('cascade');
-            $table->bigIncrements('category_id')->unsigned();
+            $table->integer('category_id')->unsigned();
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
         });
     }
